+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `toxiproxy.NewProxy` now accepts `name`, `listen addr` and `upstream addr`. (#418, @miry)
 * Replace logrus with zerolog. (#413, @miry)
 * Log HTTP requests to API server. (#413, @miry)
+* Add TimeoutHandler for the HTTP API server. (#420, @miry)
 
 # [2.4.0] - 2022-03-07